refactor: extract nav items to lib/nav-items.ts, add role-guard to detectSidebarGroup
- Create lib/nav-items.ts as single source of truth for all nav config
  (MatchContext, NavItem types; groupMatch/topLevelMatch helpers;
  USER_BASE/CLIENT_EXTRA/SPECIALIST_EXTRA/TAIL/ADMIN arrays;
  buildUserItems/itemsForGroup builders)
- SidebarNav.tsx: 585 → ~230 LOC by removing duplicated nav config,
  importing from lib/nav-items; detectSidebarGroup now accepts optional
  role param and returns "user" group when pathname is (admin-tabs) but
  user.role !== "ADMIN"
- MobileDrawer.tsx: 499 → ~240 LOC by removing all duplicated nav arrays
  and helpers, imports shared types/functions from lib/nav-items.ts;
  passes user role to detectSidebarGroup for consistent role-guard
