+++ components/layout/SidebarNav.tsx
@@ -2,24 +2,14 @@ import { useState } from "react";
 import { View, Text, Pressable, Modal, Platform } from "react-native";
 import { useRouter, usePathname, useSegments } from "expo-router";
 import { useTypedRouter } from "@/lib/navigation";
-import {
-  LayoutGrid,
-  FileText,
-  MessageCircle,
-  BarChart2,
-  Users,
-  Shield,
-  Flag,
-  Bell,
-  Settings,
-  LogOut,
-  Inbox,
-  Search,
-  type LucideIcon,
-} from "lucide-react-native";
+import { Bell, Settings, LogOut } from "lucide-react-native";
 import { colors, spacing, roleAccent, type RoleAccentKey } from "@/lib/theme";
 import { useAuth, type UserRole } from "@/contexts/AuthContext";
 import RoleBadge from "./RoleBadge";
+import {
+  type MatchContext,
+  itemsForGroup,
+} from "@/lib/nav-items";
 
 /**
  * SidebarNav — persistent left-rail navigation for authenticated routes.
@@ -33,9 +23,6 @@ import RoleBadge from "./RoleBadge";
  *   - Role-tinted background (blue client / emerald specialist / amber admin).
  *   - 3 zones: brand + role badge · primary nav · bottom identity/settings.
  *   - Active item: tint bg + left 2px accent border + bold label.
- *   - All nav items from issue GH-1285/GH-1289 spec + expanded to cover
- *     secondary destinations (Каталог специалистов, Публичные заявки,
- *     Уведомления) that were buried in AppHeader dropdown.
  *
  * Mobile (<768px): component is NOT rendered. AppShell bypasses on mobile
  * and falls back to the existing bottom-tab + burger pattern.
@@ -46,140 +33,6 @@ export type SidebarGroup =
   | "admin"
   | "main"
   | null;
-
-interface MatchContext {
-  /** Browser path (groups stripped), e.g. `/dashboard` for `/(tabs)/dashboard`. */
-  path: string;
-  /** Raw Expo-Router segments including groups, e.g. `["(tabs)", "dashboard"]`. */
-  segments: readonly string[];
-}
-
-interface NavItem {
-  label: string;
-  href: string;
-  icon: LucideIcon;
-  match: (ctx: MatchContext) => boolean;
-}
-
-// ─────────────────────────────────────────── per-role navigation maps
-
-/**
- * Match helpers.
- *
- * usePathname() strips group-parens, so `/(tabs)/dashboard` reports as
- * `/dashboard`. We use segments to know which group the user is in — that
- * disambiguates colliding paths like `/requests` that exist both as
- * `/(tabs)/requests.tsx` and `/requests/index.tsx`.
- */
-const groupMatch = (
-  ctx: MatchContext,
-  group: "(tabs)" | "(admin-tabs)",
-  leaf: string
-): boolean => {
-  if (ctx.segments[0] === group && ctx.segments[1] === leaf) return true;
-  // Safety net when segments are empty (route transitions).
-  return (
-    ctx.path.includes(`${group}/${leaf}`) ||
-    ctx.path.includes(`${group.replace(/[()]/g, "")}/${leaf}`)
-  );
-};
-
-const topLevelMatch = (ctx: MatchContext, prefix: string): boolean => {
-  // Active for `/prefix` or `/prefix/*` only when NOT inside a role-tab group
-  // (e.g. client's `/requests` screen shares URL with public `/requests`).
-  const first = ctx.segments[0] ?? "";
-  const inGroup = first.startsWith("(") && first.endsWith(")") && first !== "(tabs)";
-  if (inGroup) return false;
-  return ctx.path === prefix || ctx.path.startsWith(`${prefix}/`);
-};
-
-// Iter11 — unified USER nav (client-like items always present).
-const USER_BASE_ITEMS: NavItem[] = [
-  {
-    label: "Дашборд",
-    href: "/(tabs)",
-    icon: LayoutGrid,
-    match: (ctx) =>
-      groupMatch(ctx, "(tabs)", "index") ||
-      (ctx.segments[0] === "(tabs)" && !ctx.segments[1]),
-  },
-  {
-    label: "Мои заявки",
-    href: "/(tabs)/requests",
-    icon: FileText,
-    match: (ctx) => groupMatch(ctx, "(tabs)", "requests"),
-  },
-  {
-    label: "Сообщения",
-    href: "/(tabs)/messages",
-    icon: MessageCircle,
-    match: (ctx) => groupMatch(ctx, "(tabs)", "messages"),
-  },
-];
-
-// Client-only addition: injected after "Мои заявки" for non-specialist users.
-const USER_CLIENT_EXTRA: NavItem[] = [
-  {
-    label: "Найти специалиста",
-    href: "/specialists",
-    icon: Search,
-    match: (ctx) => topLevelMatch(ctx, "/specialists"),
-  },
-];
-
-// Specialist-only addition: appended to USER_BASE when isSpecialist=true.
-const USER_SPECIALIST_EXTRA: NavItem[] = [
-  {
-    label: "Заявки клиентов",
-    href: "/(tabs)/public-requests",
-    icon: Inbox,
-    match: (ctx) =>
-      groupMatch(ctx, "(tabs)", "public-requests") ||
-      topLevelMatch(ctx, "/requests"),
-  },
-];
-
-const USER_TAIL_ITEMS: NavItem[] = [
-  {
-    label: "Уведомления",
-    href: "/notifications",
-    icon: Bell,
-    match: (ctx) => topLevelMatch(ctx, "/notifications"),
-  },
-];
-
-const ADMIN_ITEMS: NavItem[] = [
-  {
-    label: "Дашборд",
-    href: "/(admin-tabs)/dashboard",
-    icon: BarChart2,
-    match: (ctx) => groupMatch(ctx, "(admin-tabs)", "dashboard"),
-  },
-  {
-    label: "Пользователи",
-    href: "/(admin-tabs)/users",
-    icon: Users,
-    match: (ctx) => groupMatch(ctx, "(admin-tabs)", "users"),
-  },
-  {
-    label: "Модерация",
-    href: "/(admin-tabs)/moderation",
-    icon: Shield,
-    match: (ctx) => groupMatch(ctx, "(admin-tabs)", "moderation"),
-  },
-  {
-    label: "Жалобы",
-    href: "/(admin-tabs)/complaints",
-    icon: Flag,
-    match: (ctx) => groupMatch(ctx, "(admin-tabs)", "complaints"),
-  },
-  {
-    label: "Настройки системы",
-    href: "/admin/settings",
-    icon: Settings,
-    match: (ctx) => ctx.path.startsWith("/admin/settings"),
-  },
-];
 
 // ─────────────────────────────────────────── classification helpers
 
@@ -205,13 +58,14 @@ function toAccentKey(
  *   - `useSegments()` keeps raw segments including `(tabs)` —
  *     that's the authoritative source of group membership.
  *
- * We use `segments` to detect the tab group and fall back to `pathname`
- * for top-level screens (`/requests`, `/specialists`, `/notifications`,
- * `/threads`, `/settings`) that don't live in any group.
+ * Role-guard: if pathname starts with `/(admin-tabs)` but the user is
+ * not ADMIN, we return the regular "user" group instead of "admin" so
+ * non-admins never see the admin sidebar.
  */
 export function detectSidebarGroup(
   pathname: string,
-  segments: readonly string[] = []
+  segments: readonly string[] = [],
+  role?: UserRole
 ): SidebarGroup {
   if (!pathname) return null;
 
@@ -226,11 +80,16 @@ export function detectSidebarGroup(
   const first = segments[0] ?? "";
   // Iter11 PR 3 — only (tabs) and (admin-tabs) remain; legacy groups removed.
   if (first === "(tabs)") return "user";
-  if (first === "(admin-tabs)") return "admin";
+  if (first === "(admin-tabs)") {
+    // Role-guard: only real admins get the admin sidebar.
+    return role === "ADMIN" ? "admin" : "user";
+  }
 
   // Fallback: pathname-based match for route transitions when segments
   // haven't settled yet.
-  if (pathname.includes("/admin-tabs/") || pathname.includes("(admin-tabs)")) return "admin";
+  if (pathname.includes("/admin-tabs/") || pathname.includes("(admin-tabs)")) {
+    return role === "ADMIN" ? "admin" : "user";
+  }
 
   const LEGACY_ROUTES = new Set(["/"]);
   if (LEGACY_ROUTES.has(pathname)) return null;
@@ -249,28 +108,6 @@ export function detectSidebarGroup(
   }
 
   return null;
-}
-
-function buildUserItems(isSpecialist: boolean): NavItem[] {
-  return isSpecialist
-    ? [...USER_BASE_ITEMS, ...USER_SPECIALIST_EXTRA, ...USER_TAIL_ITEMS]
-    : [...USER_BASE_ITEMS, ...USER_CLIENT_EXTRA, ...USER_TAIL_ITEMS];
-}
-
-function itemsForGroup(
-  group: SidebarGroup,
-  role: UserRole,
-  isSpecialist: boolean
-): NavItem[] {
-  if (group === "admin" || role === "ADMIN") return ADMIN_ITEMS;
-
-  switch (group) {
-    case "user":
-    case "main":
-      return buildUserItems(isSpecialist);
-    default:
-      return [];
-  }
 }
 
 export const SIDEBAR_WIDTH = 240;
